move tests into `policy` test pkg
