+++ test/integration/policy/testdata/emoji-policy.yaml
